"xcube serve2" is now "xcube serve". Removed now unused modules from old server.
