+++ xcube/webapi/places/api.py
@@ -27,6 +27,6 @@ api = Api(
     'places',
     description='xcube Places API',
     config_schema=CONFIG_SCHEMA,
-    required_apis=['_auth'],
+    required_apis=['auth'],
     create_ctx=PlacesContext
 )
